+++ XUCore/additions/NSStringGeometrics.m
@@ -16,7 +16,7 @@ int gNSStringGeometricsTypesetterBehavior = NSTypesetterLatestBehavior ;
 
 #pragma mark Measure Attributed String
 
-- (NSSize)sizeForWidth:(float)width height:(float)height {
+- (NSSize)sizeForWidth:(CGFloat)width height:(CGFloat)height {
 	NSSize answer = NSZeroSize ;
 	if ([self length] > 0) {
 		// Checking for empty string is necessary since Layout Manager will give the nominal
@@ -47,11 +47,11 @@ int gNSStringGeometricsTypesetterBehavior = NSTypesetterLatestBehavior ;
 	return answer ;
 }
 
-- (float)heightForWidth:(float)width {
+- (CGFloat)heightForWidth:(CGFloat)width {
 	return [self sizeForWidth:width height:FLT_MAX].height ;
 }
 
-- (float)widthForHeight:(float)height {
+- (CGFloat)widthForHeight:(CGFloat)height {
 	return [self sizeForWidth:FLT_MAX height:height].width ;
 }
 
@@ -62,7 +62,7 @@ int gNSStringGeometricsTypesetterBehavior = NSTypesetterLatestBehavior ;
 
 #pragma mark Given String with Attributes
 
-- (NSSize)sizeForWidth:(float)width height:(float)height attributes:(NSDictionary*)attributes {
+- (NSSize)sizeForWidth:(CGFloat)width height:(CGFloat)height attributes:(NSDictionary*)attributes {
 	NSSize answer ;
 	
 	NSAttributedString *astr = [[NSAttributedString alloc] initWithString:self attributes:attributes] ;
@@ -70,17 +70,17 @@ int gNSStringGeometricsTypesetterBehavior = NSTypesetterLatestBehavior ;
 	return answer ;
 }
 
-- (float)heightForWidth:(float)width attributes:(NSDictionary*)attributes {
+- (CGFloat)heightForWidth:(CGFloat)width attributes:(NSDictionary*)attributes {
 	return [self sizeForWidth:width height:FLT_MAX attributes:attributes].height ;
 }
 
-- (float)widthForHeight:(float)height attributes:(NSDictionary*)attributes {
+- (CGFloat)widthForHeight:(CGFloat)height attributes:(NSDictionary*)attributes {
 	return [self sizeForWidth:FLT_MAX height:height attributes:attributes].width ;
 }
 
 #pragma mark Given String with Font
 
-- (NSSize)sizeForWidth:(float)width height:(float)height font:(NSFont*)font {
+- (NSSize)sizeForWidth:(CGFloat)width height:(CGFloat)height font:(NSFont*)font {
 	NSSize answer = NSZeroSize ;
 	
 	if (font == nil) {
@@ -93,7 +93,7 @@ int gNSStringGeometricsTypesetterBehavior = NSTypesetterLatestBehavior ;
 	return answer ;
 }
 
-- (float)heightForWidth:(float)width font:(NSFont*)font {
+- (CGFloat)heightForWidth:(CGFloat)width font:(NSFont*)font {
 	NSDictionary* attributes = [NSDictionary dictionaryWithObjectsAndKeys:font, (NSString*)kCTFontAttributeName, nil];
 	NSAttributedString *attrStr = [[NSAttributedString alloc] initWithString:self attributes:attributes];
 	
@@ -123,7 +123,7 @@ int gNSStringGeometricsTypesetterBehavior = NSTypesetterLatestBehavior ;
 	return 10000.0 - origins.y + 10.0;
 }
 
-- (float)widthForHeight:(float)height font:(NSFont*)font {
+- (CGFloat)widthForHeight:(CGFloat)height font:(NSFont*)font {
 	return [self sizeForWidth:FLT_MAX height:height font:font].width ;
 }
